Move UserEditableSettings to model
